turn admin shell to a server component and clean up alternative preview component
