+++ components/admin/shell.tsx
@@ -1,93 +1,10 @@
-"use client"
-
-import { cx } from "cva"
-import {
-  CopyrightIcon,
-  GalleryHorizontalEndIcon,
-  GemIcon,
-  GlobeIcon,
-  LayoutDashboardIcon,
-  LogOutIcon,
-  ReplaceIcon,
-} from "lucide-react"
-import { signOut } from "next-auth/react"
 import type { PropsWithChildren } from "react"
-import { Nav } from "~/components/admin/nav"
-import { NavMain } from "~/components/admin/nav-main"
-import { Separator } from "~/components/admin/ui/separator"
-import { siteConfig } from "~/config/site"
-import { useIsMobile } from "~/hooks/use-mobile"
+import { Sidebar } from "~/components/admin/sidebar"
 
 export const Shell = ({ children }: PropsWithChildren) => {
-  const isMobile = useIsMobile()
-
   return (
     <div className="flex items-stretch size-full">
-      <div
-        className={cx("sticky top-0 h-dvh z-40 flex flex-col border-r", isMobile ? "w-12" : "w-48")}
-      >
-        <Nav>
-          <NavMain
-            isCollapsed={isMobile}
-            links={[
-              {
-                title: "Dashboard",
-                href: "/admin",
-                prefix: <LayoutDashboardIcon />,
-              },
-            ]}
-          />
-        </Nav>
-
-        <Separator />
-
-        <Nav>
-          <NavMain
-            isCollapsed={isMobile}
-            links={[
-              {
-                title: "Tools",
-                href: "/admin/tools",
-                prefix: <GemIcon />,
-              },
-              {
-                title: "Alternatives",
-                href: "/admin/alternatives",
-                prefix: <ReplaceIcon />,
-              },
-              {
-                title: "Categories",
-                href: "/admin/categories",
-                prefix: <GalleryHorizontalEndIcon />,
-              },
-              {
-                title: "Licenses",
-                href: "/admin/licenses",
-                prefix: <CopyrightIcon />,
-              },
-            ]}
-          />
-        </Nav>
-
-        <Nav className="mt-auto">
-          <NavMain
-            isCollapsed={isMobile}
-            links={[
-              {
-                title: "Visit Site",
-                href: siteConfig.url,
-                prefix: <GlobeIcon />,
-              },
-              {
-                title: "Sign Out",
-                href: "#",
-                onClick: () => signOut(),
-                prefix: <LogOutIcon />,
-              },
-            ]}
-          />
-        </Nav>
-      </div>
+      <Sidebar />
 
       <div className="grid grid-cols-1 content-start gap-4 p-4 flex-1 sm:px-6">{children}</div>
     </div>
